Add & fix the kdoc of the `model`s
