+++ clashJ/src/main/kotlin/com/clashj/model/clan/Clan.kt
@@ -11,37 +11,38 @@ import com.clashj.model.clan.component.ClanMember
 import com.clashj.model.clan.component.WarFrequency
 
 /**
- * Represents detailed information about a clan in a game.
+ * Represents the *Clan* model of the Clash of Clans API.
+ * The Clan data class stores detailed information about a clan in the game.
  *
- * @property warLeague The league of the clan in wars.
- * @property capitalLeague The league of the clan's capital.
- * @property memberList The list of clan members.
+ * @property warLeague The league information for clan wars.
+ * @property capitalLeague The league information for the clan's capital.
+ * @property memberList A list of [ClanMember] objects representing the clan's members.
  * @property tag The tag of the clan.
- * @property warFrequency The frequency of wars for the clan.
+ * @property warFrequency The frequency of clan wars.
  * @property clanLevel The level of the clan.
- * @property warWinStreak The current war win streak of the clan.
- * @property warWins The total number of war wins for the clan.
- * @property warTies The total number of war ties for the clan.
- * @property warLosses The total number of war losses for the clan.
- * @property clanPoints The points of the clan.
- * @property requiredTownhallLevel The required town hall level to join the clan.
- * @property chatLanguage The language used in the clan's chat.
- * @property isFamilyFriendly Indicates whether the clan is family-friendly.
- * @property requiredTrophies The required trophies to join the clan.
- * @property requiredBuilderBaseTrophies The required builder base trophies to join the clan.
- * @property requiredVersusTrophies The required versus trophies to join the clan.
- * @property isWarLogPublic Indicates whether the clan's war log is public.
- * @property clanBuilderBasePoints The builder base points of the clan.
- * @property clanVersusPoints The versus points of the clan.
- * @property clanCapitalPoints The capital points of the clan.
- * @property labels The list of labels associated with the clan.
+ * @property warWinStreak The current winning streak in clan wars.
+ * @property warWins The total number of war wins.
+ * @property warTies The total number of war ties.
+ * @property warLosses The total number of war losses.
+ * @property clanPoints The total points earned by the clan.
+ * @property requiredTownhallLevel The minimum required town hall level to join the clan.
+ * @property chatLanguage The language used for clan chat.
+ * @property isFamilyFriendly A flag indicating if the clan is family-friendly.
+ * @property requiredTrophies The minimum required trophies to join the clan.
+ * @property requiredBuilderBaseTrophies The minimum required builder base trophies to join the clan.
+ * @property requiredVersusTrophies The minimum required versus trophies to join the clan.
+ * @property isWarLogPublic A flag indicating if the clan's war log is public.
+ * @property clanBuilderBasePoints The builder base points earned by the clan.
+ * @property clanVersusPoints The versus points earned by the clan.
+ * @property clanCapitalPoints The points earned by the clan's capital.
+ * @property labels A list of [Label] objects categorizing the clan.
  * @property name The name of the clan.
  * @property location The location of the clan.
- * @property type The type of the clan (Open, Invite Only, Closed).
- * @property members The total number of members in the clan.
+ * @property type The type of the clan (open, invite-only, closed).
+ * @property members The total number of clan members.
  * @property description The description of the clan.
- * @property clanCapital The clan's capital information.
- * @property badgeUrls The URLs to the clan's badges (small, medium, large).
+ * @property clanCapital The information about the clan's capital.
+ * @property badgeUrls The badge URLs representing the clan's badge in different sizes.
  */
 data class Clan(
     val warLeague: SimpleLeague,
