+++ clashJ/src/main/kotlin/com/clashj/model/clan/ClanMemberList.kt
@@ -4,10 +4,10 @@ import com.clashj.model.common.Paging
 import com.clashj.model.clan.component.ClanMember
 
 /**
- * Represents the Clan War Log.
+ * Represents the *ClanMemberList* model of the Clash of Clans API.
  *
- * @property items List of [ClanMember].
- * @property paging Paging information for pagination. Can be null if there is no pagination.
+ * @property items A list of [ClanMember] objects representing the clan members in the list.
+ * @property paging The pagination information for navigating through the list (optional).
  */
 data class ClanMemberList(
     val items: List<ClanMember>,
