+++ clashJ/src/main/kotlin/com/clashj/model/clan/ClanWarLog.kt
@@ -5,10 +5,10 @@ import com.clashj.model.clan.component.WarClan
 import com.clashj.model.clan.component.WarResult
 
 /**
- * Represents the Clan War Log.
+ * Represents the *ClanWarLog* model of the Clash of Clans API.
  *
- * @property items List of [ClanWarLogEntry].
- * @property paging Paging information for pagination. Can be null if there is no pagination.
+ * @property items A list of [ClanWarLogEntry] objects representing the clan war log entries in the list.
+ * @property paging The pagination information for navigating through the list (optional).
  */
 data class ClanWarLog(
     val items: List<ClanWarLogEntry>,
@@ -16,14 +16,15 @@ data class ClanWarLog(
 )
 
 /**
- * Represents an entry in the Clan War Log.
+ * Represents the *ClanWarLogEntry* model of the Clash of Clans API.
+ * The ClanWarLogEntry data class stores information about a single clan war log entry.
  *
- * @property clan The clan information for the entry.
- * @property teamSize The size of the team in the war.
+ * @property clan The information about the clan participating in the war.
+ * @property teamSize The size of the team participating in the war.
  * @property attacksPerMember The number of attacks allowed per member in the war.
- * @property opponent The information of the opponent clan.
- * @property endTime The end time of the war in string format.
- * @property result The result of the war for the clan.
+ * @property opponent The information about the opponent clan in the war.
+ * @property endTime The end time of the war in ISO 8601 date-time format.
+ * @property result The result of the clan war (win, lose, tie).
  */
 data class ClanWarLogEntry(
     val clan: WarClan,
