+++ clashJ/src/main/kotlin/com/clashj/model/clan/component/ClanCapital.kt
@@ -1,10 +1,11 @@
 package com.clashj.model.clan.component
 
 /**
- * Represents a Clash of Clans API ClanCapital.
+ * Represents the *ClanCapital* model of the Clash of Clans API.
+ * The ClanCapital data class stores information about a clan's capital and its districts.
  *
  * @property capitalHallLevel The hall level of the clan's capital.
- * @property districts The list of [District]s belonging to the capital.
+ * @property districts A list of [District] objects containing information about the capital's districts.
  */
 data class ClanCapital(
     val capitalHallLevel: Int,
@@ -12,10 +13,11 @@ data class ClanCapital(
 )
 
 /**
- * Represents a district belonging to a clan's capital.
+ * Represents the *District* model of the Clash of Clans API.
+ * The District data class stores information about a district within a clan's capital.
  *
  * @property name The name of the district.
- * @property id The ID of the district.
+ * @property id The unique identifier for the district.
  * @property districtHallLevel The hall level of the district.
  */
 data class District(
