+++ clashJ/src/main/kotlin/com/clashj/model/clan/component/ClanMember.kt
@@ -5,22 +5,22 @@ import com.clashj.model.common.PlayerHouse
 import com.clashj.model.league.SimpleLeague
 
 /**
- * Represents a member of a clan in the game.
+ * Represents the *ClanMember* model of the Clash of Clans API.
  *
- * @property league The league of the clan member.
- * @property builderBaseLeague The league of the clan member in Builder Base.
- * @property playerHouse The player house elements of the clan member.
- * @property tag The unique tag of the clan member.
- * @property versusTrophies The number of versus trophies of the clan member.
+ * @property league The league information of the clan member in the main village.
+ * @property builderBaseLeague The league information of the clan member in the builder base village.
+ * @property playerHouse The player house information of the clan member.
+ * @property tag The tag of the clan member.
+ * @property versusTrophies The number of versus trophies earned by the clan member.
  * @property name The name of the clan member.
- * @property role The role of the clan member in the clan.
+ * @property role The role of the clan member within the clan.
  * @property expLevel The experience level of the clan member.
- * @property clanRank The rank of the clan member in the clan.
- * @property previousClanRank The previous rank of the clan member in the clan.
+ * @property clanRank The rank of the clan member within the clan.
+ * @property previousClanRank The previous rank of the clan member within the clan.
  * @property donations The number of troops donated by the clan member.
- * @property donationsReceived The number of troops received by the clan member.
- * @property trophies The number of trophies of the clan member.
- * @property builderBaseTrophies The number of Builder Base trophies of the clan member.
+ * @property donationsReceived The number of troops received by the clan member from other clan members.
+ * @property trophies The number of trophies earned by the clan member in the main village.
+ * @property builderBaseTrophies The number of trophies earned by the clan member in the builder base village.
  */
 data class ClanMember(
     val league: League,
