+++ clashJ/src/main/kotlin/com/clashj/model/common/Paging.kt
@@ -1,22 +1,23 @@
 package com.clashj.model.common
 
 /**
- * Represents pagination information.
+ * Represents the *Paging* model of the Clash of Clans API.
  *
- * @property cursors The cursors for navigating through the data.
+ * @property cursors The cursors for pagination, containing references to the `before` and `after` elements in a list.
+ * It provides navigation to the previous and next pages of results.
  */
 data class Paging(
     val cursors: Cursors
 )
 
 /**
- * Represents cursors for pagination.
+ * Represents the *Cursors* model of the Clash of Clans API.
+ * Cursors are used for pagination, providing references to the "before" and "after" elements in a list.
  *
- * @property before The cursor for returning items that occur before this marker.
- * Can be null if there is no before marker.
- *
- * @property after The cursor for returning items that occur after this marker.
- * Can be null if there is no after marker.
+ * @property before The cursor pointing to the element before the current page in the list.
+ * It is null for the first page of results.
+ * @property after The cursor pointing to the element after the current page in the list.
+ * It is null for the last page of results.
  */
 data class Cursors(
     val before: String?,
